backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
colorama
openpyxl
sqlalchemy
python-jose[cryptography]
pyinstaller
